refactor: atomWithDepositSettings
