+++ src/features/phat-contract/hooks/useContractExecutor.ts
@@ -1,5 +1,8 @@
 import type {Bytes} from '@polkadot/types-codec'
 import type {ContractOptions} from '@polkadot/api-contract/types'
+import type { u64 } from '@polkadot/types'
+import type { BN } from '@polkadot/util'
+import type { DepositSettings } from '../atomsWithDepositSettings'
 
 import { useState, useCallback } from 'react'
 import { atom, useAtomValue, useSetAtom } from "jotai"
@@ -30,6 +33,7 @@ import {
   currentWorkerIdAtom,
 } from '../atoms'
 
+
 interface InkResponse {
   nonce: string
   result: {
@@ -39,17 +43,17 @@ interface InkResponse {
   }
 }
 
-export interface DepositSettings {
-  autoDeposit: boolean
-  gasLimit?: number | null
-  storageDepositLimit?: number | null
-}
-
 const debug = createLogger('chain', 'debug')
+
+
+interface EstimateGasResult {
+  gasLimit: u64
+  storageDepositLimit: BN | null
+}
 
 async function estimateGas(contract: ContractPromise, method: string, cert: CertificateData, args: unknown[]) {
   const { gasRequired, storageDeposit } = await contract.query[method](cert as any, {}, ...args)
-  const options = {
+  const options: EstimateGasResult = {
       gasLimit: (gasRequired as any).refTime,
       storageDepositLimit: storageDeposit.isCharge ? storageDeposit.asCharge : null
   }
@@ -104,7 +108,6 @@ export const estimateGasAtom = atom(async get => {
     selectedMethodSpec!.args
   )
   const txConf = await estimateGas(contractInstance, txMethods[selectedMethodSpec!.label], cert, args);
-  console.log('useContractEstimeateGas', txConf.gasLimit.toHuman(), txConf.storageDepositLimit?.toHuman())
   return txConf
 })
